addded cancel button and its functionality

+++ rent/views.py
@@ -147,21 +147,24 @@ class AddBuilding(BaseView):
         return render(request,'building_action.html',context)
     
     def post(self,request):
-        building_name = request.POST.get('building_name')
-        building_address = request.POST.get('building_address')
-        room_count = request.POST.get('room_count')
-        building_type = request.POST.get('building_type')
-        remarks = request.POST.get('remarks')
-        Building.objects.create(
-            building_name=building_name,
-            building_address=building_address,
-            building_type=building_type,
-            number_of_rooms=room_count,
-            is_active=True,
-            remarks = remarks
-        )
-        messages.success(request,"Building added successfully")
-        return redirect('rent:home')
+        if 'cancel' in request.POST:
+            return redirect('rent:select-building')
+        elif 'add' in request.POST:
+            building_name = request.POST.get('building_name')
+            building_address = request.POST.get('building_address')
+            room_count = request.POST.get('room_count')
+            building_type = request.POST.get('building_type')
+            remarks = request.POST.get('remarks')
+            Building.objects.create(
+                building_name=building_name,
+                building_address=building_address,
+                building_type=building_type,
+                number_of_rooms=room_count,
+                is_active=True,
+                remarks = remarks
+            )
+            messages.success(request,"Building added successfully")
+            return redirect('rent:home')
 
 class UpdateBuilding(View):
     def get(self, request, building_name):
@@ -176,16 +179,19 @@ class UpdateBuilding(View):
         return render(request, 'building_action.html', context)
 
     def post(self, request, building_name):
-        building = get_object_or_404(Building, uid=building_name)
-        building.building_name = request.POST.get('building_name')
-        building.building_address = request.POST.get('building_address')
-        building.number_of_rooms = request.POST.get('room_count')
-        building.building_type = request.POST.get('building_type')
-        building.remarks = request.POST.get('remarks')
-        building.save()
-        
-        messages.success(request, "Building has been updated successfully")
-        return redirect('rent:home')
+        if 'cancel' in request.POST:
+            return redirect('rent:select-building')
+        elif 'update' in request.POST:
+            building = get_object_or_404(Building, uid=building_name)
+            building.building_name = request.POST.get('building_name')
+            building.building_address = request.POST.get('building_address')
+            building.number_of_rooms = request.POST.get('room_count')
+            building.building_type = request.POST.get('building_type')
+            building.remarks = request.POST.get('remarks')
+            building.save()
+            
+            messages.success(request, "Building has been updated successfully")
+            return redirect('rent:home')
 
 class DeleteBuilding(View):
     def get(self, request, building_name):
@@ -200,11 +206,13 @@ class DeleteBuilding(View):
         return render(request, 'building_action.html', context)
 
     def post(self, request, building_name):
-        building = get_object_or_404(Building, uid=building_name)
-        building.delete()
-        
-        messages.success(request, "Building has be removed successfully")
-        return redirect('rent:home')        
+        if 'cancel' in request.POST:
+            return redirect('rent:select-building')
+        elif 'delete' in request.POST:
+            building = get_object_or_404(Building, uid=building_name)
+            building.delete()
+            messages.success(request, "Building has be removed successfully")
+            return redirect('rent:home')        
 
 
 class ViewRentHistory(BaseView):
